refactor: abstract filter components
